fix(security): model token version without signed-int truncation

+++ app/src/main/java/com/aqua/aqualight/data/devices/runtime/modules/security/DeviceSecurityModels.kt
@@ -40,7 +40,7 @@ data class DeviceSecurityCredentialStorage(
     val storedPlaintext: Boolean,
     val tokenFormat: String,
     val tokenHexLength: Int,
-    val tokenVersion: Int?,
+    val tokenVersion: Long?,
     val pairedAtMs: Long?,
     val lastRotatedAtMs: Long?,
     val provisioningTokenPending: Boolean?
